refactor: #251 TagList 컴포넌트를 components 폴더로 이동
